configure phpunit to dont break 8.1 tests
