+++ composer.json
@@ -53,7 +53,8 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Backpack\\CRUD\\Tests\\": "tests"
+            "Backpack\\CRUD\\Tests\\": "tests",
+            "Backpack\\CRUD\\Tests81\\": "tests81"
         },
         "files": ["src/helpers.php"],
         "exclude-from-classmap": [
