+++ tests81/Unit/Models/Enums/StatusEnum.php
@@ -1,6 +1,9 @@
 <?php
 
-namespace Backpack\CRUD\Tests\Unit\Models\Enums;
+namespace Backpack\CRUD\Tests81\Unit\Models\Enums;
+
+if(version_compare(PHP_VERSION,'8.1', '>='))
+{
 
 enum StatusEnum: string
 {
@@ -11,4 +14,6 @@ enum StatusEnum: string
     {
         return array_combine(array_column(self::cases(), 'name'), array_column(self::cases(), 'value'));
     }
+}
+
 }
